Refactor Module Progress Component
Refactors the module progress bars into their own Component.

* Create ModuleProgressComponent
* Remove code from App.jsx and replace with Component reference.
