+++ src/App.jsx
@@ -8,6 +8,8 @@ import GUIDE_CONTENT from './data/guide-content'
 import MODEL_RECOMMENDATIONS from './data/model-recommendations'
 import QUESTIONNAIRES from './data/questionnaires'
 
+import ModuleProgressComponent from './components/component-module-progress'
+
 import './index.css'
 
 // Calculate score and recommendation functions
@@ -306,31 +308,7 @@ function App() {
                 <hr className="my-4" />
                 
                 <div className="mt-4">
-                  <h3 className="font-medium mb-2">Module Progress</h3>
-                  {Object.entries(AI_MODULES_INFO).map(([key, value]) => {
-                    const moduleAnswers = answers[key];
-                    const questionsAnswered = Object.keys(moduleAnswers).length;
-                    const totalQuestions = QUESTIONNAIRES[key].length;
-                    const progressPercentage = Math.round((questionsAnswered / totalQuestions) * 100);
-                    
-                    return (
-                      <div key={key} className="mb-3">
-                        <div className="flex justify-between text-sm mb-1">
-                          <span className="text-xs">{value.title}</span>
-                          <span className="text-xs">{questionsAnswered}/{totalQuestions}</span>
-                        </div>
-                        <div className="w-full bg-gray-200 rounded-full h-2">
-                          <div 
-                            className="h-2 rounded-full" 
-                            style={{
-                              width: `${progressPercentage}%`,
-                              backgroundColor: value.color
-                            }}
-                          ></div>
-                        </div>
-                      </div>
-                    );
-                  })}
+                  <ModuleProgressComponent AI_MODULES_INFO={AI_MODULES_INFO} answers={answers} QUESTIONNAIRES={QUESTIONNAIRES} />
                 </div>
                 
                 <div className="mt-6 bg-blue-50 p-3 rounded-lg">
